+++ azurerm/resourcegroups.py
@@ -1,7 +1,7 @@
 '''resourcegroups.py - azurerm functions for Resource Groups.'''
 import json
-from .restfns import do_delete, do_get, do_put
-from .settings import get_rm_endpoint, BASE_API
+from .restfns import do_delete, do_get, do_post, do_put
+from .settings import get_rm_endpoint, RESOURCE_API
 
 
 def create_resource_group(access_token, subscription_id, rgname, location):
@@ -19,7 +19,7 @@ def create_resource_group(access_token, subscription_id, rgname, location):
     endpoint = ''.join([get_rm_endpoint(),
                         '/subscriptions/', subscription_id,
                         '/resourcegroups/', rgname,
-                        '?api-version=', BASE_API])
+                        '?api-version=', RESOURCE_API])
     rg_body = {'location': location}
     body = json.dumps(rg_body)
     return do_put(endpoint, body, access_token)
@@ -39,8 +39,29 @@ def delete_resource_group(access_token, subscription_id, rgname):
     endpoint = ''.join([get_rm_endpoint(),
                         '/subscriptions/', subscription_id,
                         '/resourcegroups/', rgname,
-                        '?api-version=', BASE_API])
+                        '?api-version=', RESOURCE_API])
     return do_delete(endpoint, access_token)
+
+
+def export_template(access_token, subscription_id, rgname):
+    '''Capture the specified resource group as a template
+
+    Args:
+        access_token (str): A valid Azure authentication token.
+        subscription_id (str): Azure subscription id.
+        rgname (str): Azure resource group name.
+
+    Returns:
+        HTTP response. JSON body.
+    '''
+    endpoint = ''.join([get_rm_endpoint(),
+                        '/subscriptions/', subscription_id,
+                        '/resourcegroups/', rgname,
+                        '/exportTemplate',
+                        '?api-version=', RESOURCE_API])
+    rg_body = {'options':'IncludeParameterDefaultValue', 'resources':['*']}
+    body = json.dumps(rg_body)
+    return do_post(endpoint, body, access_token)
 
 
 def get_resource_group(access_token, subscription_id, rgname):
@@ -52,12 +73,30 @@ def get_resource_group(access_token, subscription_id, rgname):
         rgname (str): Azure resource group name.
 
     Returns:
-        HTTP response.
+        HTTP response. JSON body.
     '''
     endpoint = ''.join([get_rm_endpoint(),
                         '/subscriptions/', subscription_id,
                         '/resourceGroups/', rgname,
-                        '?api-version=', BASE_API])
+                        '?api-version=', RESOURCE_API])
+    return do_get(endpoint, access_token)
+
+
+def get_resource_group_resources(access_token, subscription_id, rgname):
+    '''Get the resources in the named resource group.
+
+    Args:
+        access_token (str): A valid Azure authentication token.
+        subscription_id (str): Azure subscription id.
+        rgname (str): Azure resource group name.
+
+    Returns:
+        HTTP response. JSON body.
+    '''
+    endpoint = ''.join([get_rm_endpoint(),
+                        '/subscriptions/', subscription_id,
+                        '/resourceGroups/', rgname,
+                        '/resources?api-version=', RESOURCE_API])
     return do_get(endpoint, access_token)
 
 
@@ -74,5 +113,5 @@ def list_resource_groups(access_token, subscription_id):
     endpoint = ''.join([get_rm_endpoint(),
                         '/subscriptions/', subscription_id,
                         '/resourceGroups/',
-                        '?api-version=', BASE_API])
+                        '?api-version=', RESOURCE_API])
     return do_get(endpoint, access_token)