+++ test/resource_groups_test.py
@@ -5,7 +5,10 @@ import sys
 import unittest
 from haikunator import Haikunator
 import json
+import time
+
 import azurerm
+
 
 class TestAzurermPy(unittest.TestCase):
 
@@ -21,7 +24,8 @@ class TestAzurermPy(unittest.TestCase):
         app_id = configData['appId']
         app_secret = configData['appSecret']
         self.subscription_id = configData['subscriptionId']
-        self.access_token = azurerm.get_access_token(tenant_id, app_id, app_secret)
+        self.access_token = azurerm.get_access_token(
+            tenant_id, app_id, app_secret)
         self.location = configData['location']
         h = Haikunator()
         self.rgname = h.haikunate()
@@ -32,26 +36,41 @@ class TestAzurermPy(unittest.TestCase):
     def test_resource_groups(self):
         # create resource group
         print('Creating resource group: ' + self.rgname)
-        response = azurerm.create_resource_group(self.access_token, self.subscription_id, \
-            self.rgname, self.location)
+        response = azurerm.create_resource_group(self.access_token, self.subscription_id,
+                                                 self.rgname, self.location)
         self.assertEqual(response.status_code, 201)
 
         # get resource group
         print('Getting resource group: ' + self.rgname)
-        response = azurerm.get_resource_group(self.access_token, self.subscription_id, self.rgname)
+        response = azurerm.get_resource_group(
+            self.access_token, self.subscription_id, self.rgname)
         self.assertEqual(response['name'], self.rgname)
+
+        # export resource group
+        print('Exporting resource group: ' + self.rgname)
+        response = azurerm.export_template(
+            self.access_token, self.subscription_id, self.rgname)
+        self.assertEqual(response.status_code, 200)
+
+        # get resource group resources
+        print('Getting resources for resource group: ' + self.rgname)
+        response = azurerm.get_resource_group_resources(self.access_token, self.subscription_id,
+                                                        self.rgname)
+        #print(json.dumps(response, sort_keys=False, indent=2, separators=(',', ': ')))
+        self.assertTrue('value' in response)
 
         # list resource groups
         print('List resource groups: ' + self.rgname)
-        response = azurerm.list_resource_groups(self.access_token, self.subscription_id)
+        response = azurerm.list_resource_groups(
+            self.access_token, self.subscription_id)
         self.assertTrue('value' in response)
 
         # delete resource group
         print('Deleting resource group: ' + self.rgname)
-        response = azurerm.delete_resource_group(self.access_token, self.subscription_id, self.rgname)
+        response = azurerm.delete_resource_group(
+            self.access_token, self.subscription_id, self.rgname)
         self.assertEqual(response.status_code, 202)
 
 
 if __name__ == '__main__':
     unittest.main()
-